backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
pillow==10.0.1
pytesseract==0.3.10
PyPDF2==3.0.1
python-docx==1.1.0
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
